requirements.txt
Flask
PyPDF2
python-docx
joblib
scikit-learn
